Adds -tags support
This commit is a patch based on https://github.com/tools/godep/pull/117
that adds support for -tags to godep.

It is REQUIRED in order to save the deps for a different platform on
your dev machine; in our case it is to save off the heroku deps
in addition to our application deps in order to pull in and build the
goose migration tool remotely and have it available on heroku.